+++ src/rust/async/waker.h
@@ -199,20 +199,4 @@ private:
   FuturePollEvent& futurePollEvent;
 };
 
-// =======================================================================================
-// OptionWaker
-
-// An opaque Rust type defined in lib.rs, and thus in lib.rs.h. lib.rs.h depends on our C++ headers,
-// including waker.h (the file you're currently reading), so we forward-declare OptionWaker here for
-// use in the C++ headers.
-//
-// This class is a wrapper around an arbitrary `std::task::Waker`. It has one function on it:
-// `wake_by_ref()`. RustPromiseAwaiter calls it when the KJ promise it is currently waiting on
-// becomes ready. In this way, Rust is able to await KJ promises using any Waker of their choosing,
-// not just a LazyArcWaker.
-struct OptionWaker;
-
-// Wrapper around an arbitrary `&std::task::Waker`.
-struct WakerRef;
-
 }  // namespace workerd::rust::async
