+++ src/rust/async/waker.rs
@@ -93,41 +93,3 @@ pub fn try_into_kj_waker_ptr<'a>(waker: &Waker) -> *const KjWaker {
         std::ptr::null() as *const KjWaker
     }
 }
-
-// =======================================================================================
-// OptionWaker
-//
-// This is a wrapper around `std::task::Waker`, exposed to C++. We use it in `RustPromiseAwaiter`
-// to allow KJ promises to be awaited using opaque Wakers implemented in Rust.
-
-pub struct OptionWaker {
-    inner: Option<Waker>,
-}
-
-impl OptionWaker {
-    pub fn empty() -> OptionWaker {
-        OptionWaker { inner: None }
-    }
-
-    pub fn set(&mut self, waker: &crate::WakerRef) {
-        if let Some(w) = &mut self.inner {
-            w.clone_from(waker.0);
-        } else {
-            self.inner = Some(waker.0.clone());
-        }
-    }
-
-    pub fn set_none(&mut self) {
-        self.inner = None;
-    }
-
-    pub fn wake(&mut self) {
-        self.inner
-            .take()
-            .expect(
-                "OptionWaker::set() should be called before RustPromiseAwaiter::poll(); \
-                OptionWaker::wake() should be called at most once after poll()",
-            )
-            .wake();
-    }
-}
